Ensure testharness makes missing doctype name null
This change ensures the token handler used by the testharness sets the
doctype name to null when the doctype name is missing in the input source.

Otherwise, without this change, the testharness token handler sets the
doctype name to the empty string — which causes us to fail 9 tests in
the html5lib-tests suite.

Relates to https://github.com/validator/htmlparser/issues/35